Added an API utils library for calling the API in the fashion we like for ingestion. Tested with the run of main and everything works fine.
